refactor: split users controller into account/org_roles package per ADR-0008 (#2253)
