+++ src/synthorg/api/controllers/users/account.py
@@ -1,17 +1,23 @@
-"""User management controller -- CEO-only CRUD for human users."""
+# module-kind: controller
+"""User account controller -- CEO-only CRUD for human users."""
 
 import uuid
 from datetime import UTC, datetime
-from typing import Final, LiteralString
+from typing import Final
 
 from litestar import Controller, Request, delete, get, patch, post
 from litestar.datastructures import State
 from litestar.status_codes import HTTP_204_NO_CONTENT
-from pydantic import AwareDatetime, BaseModel, ConfigDict, Field
+from pydantic import BaseModel, ConfigDict, Field
 
 from synthorg._core.features import require_service
 from synthorg.api.api_core_state import ApiCoreStateSlice
-from synthorg.api.auth.user_service import UserService
+from synthorg.api.controllers.users._shared import (
+    UserResponse,
+    _get_user_or_404,
+    _service,
+    _to_response,
+)
 from synthorg.api.cursor import decode_keyset_cursor
 from synthorg.api.dto import ApiResponse, PaginatedResponse
 from synthorg.api.guards import require_ceo
@@ -21,14 +27,12 @@ from synthorg.api.pagination import (
     cursor_secret_of,
     encode_keyset_meta,
 )
-from synthorg.api.path_params import PathId, PathName
+from synthorg.api.path_params import PathId
 from synthorg.api.rate_limits import per_op_rate_limit_from_policy
-from synthorg.api.responses import require_resource_or_404
 from synthorg.api.state import AppState
 from synthorg.core.auth.config import AuthConfig
-from synthorg.core.auth.models import AuthenticatedUser, OrgRole, User
+from synthorg.core.auth.models import AuthenticatedUser, User
 from synthorg.core.auth.roles import HumanRole
-from synthorg.core.collections import dedupe_preserving_order
 from synthorg.core.domain_errors import ConflictError, NotFoundError, ValidationError
 from synthorg.core.persistence_errors import ConstraintViolationError, QueryError
 from synthorg.core.types import NotBlankStr
@@ -39,51 +43,21 @@ from synthorg.observability.events.api import (
     API_USER_SAVE_FAILED,
     API_VALIDATION_FAILED,
 )
-from synthorg.observability.events.security import (
-    SECURITY_PERMISSION_GRANTED,
-    SECURITY_PERMISSION_REVOKED,
-)
 from synthorg.persistence.constraint_tokens import (
     IDX_SINGLE_CEO,
     LAST_CEO_TRIGGER,
     LAST_OWNER_TRIGGER,
     USERS_USERNAME_UNIQUE,
 )
-from synthorg.persistence.state import persistence_of
 
 logger = get_logger(__name__)
 _DEFAULT_LIMIT: Final[int] = 50
-
-
-def _service(state: State) -> UserService:
-    """Build the per-request :class:`UserService`.
-
-    Threads the refresh-token repo so ``delete()`` can explicitly
-    revoke outstanding refresh tokens before the DB delete as
-    defense-in-depth (CFG-1 audit). Sessions, api_keys, and
-    refresh_tokens are all also removed by the schema's
-    ``ON DELETE CASCADE`` on ``user_id`` when the user row goes
-    away -- the explicit revocation runs first so tokens stop
-    minting access tokens immediately.
-
-    Returns:
-        ``UserService`` instance.
-    """
-    persistence = persistence_of(state.app_state)
-    return UserService(
-        repo=persistence.users,
-        refresh_tokens=persistence.refresh_tokens,
-    )
-
 
 # Derive from AuthConfig default to prevent silent divergence.
 _MIN_PASSWORD_LENGTH: int = AuthConfig.model_fields["min_password_length"].default
 
 # Roles that cannot be assigned via the user management API.
 _FORBIDDEN_ROLES: frozenset[HumanRole] = frozenset({HumanRole.SYSTEM})
-
-
-# -- Request / Response DTOs -------------------------------------------
 
 
 class CreateUserRequest(BaseModel):
@@ -104,51 +78,6 @@ class UpdateUserRoleRequest(BaseModel):
     role: HumanRole
 
 
-class GrantOrgRoleRequest(BaseModel):
-    """Request body for granting an org-level role."""
-
-    model_config = ConfigDict(frozen=True, allow_inf_nan=False, extra="forbid")
-
-    role: OrgRole
-    scoped_departments: tuple[NotBlankStr, ...] = ()
-
-
-class UserResponse(BaseModel):
-    """Public user representation (no password hash)."""
-
-    model_config = ConfigDict(frozen=True, allow_inf_nan=False, extra="forbid")
-
-    id: NotBlankStr
-    username: NotBlankStr
-    role: HumanRole
-    must_change_password: bool
-    org_roles: tuple[str, ...] = ()
-    scoped_departments: tuple[str, ...] = ()
-    created_at: AwareDatetime
-    updated_at: AwareDatetime
-
-
-def _to_response(user: User) -> UserResponse:
-    """Map a ``User`` domain model to the public ``UserResponse`` DTO.
-
-    Returns:
-        ``UserResponse`` instance.
-    """
-    return UserResponse(
-        id=user.id,
-        username=user.username,
-        role=user.role,
-        must_change_password=user.must_change_password,
-        org_roles=tuple(r.value for r in user.org_roles),
-        scoped_departments=user.scoped_departments,
-        created_at=user.created_at,
-        updated_at=user.updated_at,
-    )
-
-
-# -- Validation helpers ------------------------------------------------
-
-
 def _validate_assignable_role(role: HumanRole) -> None:
     """Reject roles that cannot be assigned via the API.
 
@@ -159,34 +88,6 @@ def _validate_assignable_role(role: HumanRole) -> None:
         msg = f"Cannot assign role: {role.value}"
         logger.warning(API_VALIDATION_FAILED, reason=msg)
         raise ValidationError(msg)
-
-
-async def _get_user_or_404(
-    service: UserService,
-    user_id: str,
-    *,
-    operation: LiteralString,
-) -> User:
-    """Fetch a user by ID, raising NotFoundError if missing.
-
-    ``operation`` discriminates the failing endpoint in the audit log
-    (``"read"`` / ``"update_user_role"`` / ``"delete_user"`` /
-    ``"grant_org_role"`` / ``"revoke_org_role"``) so the not-found
-    emissions are not all stamped as ``"read"``.
-
-    Returns:
-        ``User`` instance.
-    """
-    return require_resource_or_404(
-        await service.get(NotBlankStr(user_id)),
-        resource_type="User",
-        identifier=user_id,
-        log_event=API_RESOURCE_NOT_FOUND,
-        operation=operation,
-    )
-
-
-# -- Controller --------------------------------------------------------
 
 
 class UserController(Controller):
@@ -507,199 +408,3 @@ class UserController(Controller):
             msg = f"User not found: {user_id}"
             logger.warning(API_RESOURCE_NOT_FOUND, reason=msg)
             raise NotFoundError(msg)
-
-    # -- Org role grant/revoke -------------------------------------------
-
-    @post(
-        "/{user_id:str}/org-roles",
-        status_code=201,
-        guards=[
-            per_op_rate_limit_from_policy("users.grant_org_role", key="user"),
-        ],
-    )
-    async def grant_org_role(
-        self,
-        state: State,
-        user_id: PathId,
-        data: GrantOrgRoleRequest,
-    ) -> ApiResponse[UserResponse]:
-        """Grant an org-level role to a user.
-
-        Args:
-            state: Application state.
-            user_id: Target user identifier.
-            data: Role grant payload.
-
-        Returns:
-            Updated user response (HTTP 201).
-
-        Raises:
-            NotFoundError: If the user is not found.
-            ConflictError: If the user already has the role.
-            ValidationError: If department_admin without departments.
-            ConstraintViolationError: Raised on the corresponding failure path.
-            QueryError: Raised on the corresponding failure path.
-        """
-        service = _service(state)
-        user = await _get_user_or_404(service, user_id, operation="grant_org_role")
-
-        if user.role == HumanRole.SYSTEM:
-            msg = "Cannot assign org roles to the system user"
-            logger.warning(API_VALIDATION_FAILED, reason=msg)
-            raise ValidationError(msg)
-
-        existing_roles = set(user.org_roles)
-        if data.role in existing_roles:
-            msg = f"User already has role: {data.role.value}"
-            logger.warning(API_RESOURCE_CONFLICT, reason=msg)
-            raise ConflictError(msg)
-
-        if data.role == OrgRole.DEPARTMENT_ADMIN and not data.scoped_departments:
-            msg = "department_admin role requires scoped_departments"
-            logger.warning(API_VALIDATION_FAILED, reason=msg)
-            raise ValidationError(msg)
-        if data.role != OrgRole.DEPARTMENT_ADMIN and data.scoped_departments:
-            msg = "scoped_departments can only be set for department_admin"
-            logger.warning(API_VALIDATION_FAILED, reason=msg)
-            raise ValidationError(msg)
-
-        new_roles = (*user.org_roles, data.role)
-        new_scoped = (
-            tuple(
-                sorted(
-                    dedupe_preserving_order(
-                        [*user.scoped_departments, *data.scoped_departments],
-                    ),
-                )
-            )
-            if data.role == OrgRole.DEPARTMENT_ADMIN
-            else user.scoped_departments
-        )
-        now = datetime.now(UTC)
-        updated = user.model_copy(
-            update={
-                "org_roles": new_roles,
-                "scoped_departments": new_scoped,
-                "updated_at": now,
-            },
-        )
-        try:
-            await service.save_update(
-                updated,
-                intent="grant_org_role",
-                granted_org_role=data.role.value,
-            )
-        except ConstraintViolationError as exc:
-            if exc.constraint == LAST_OWNER_TRIGGER:
-                msg = "Cannot modify the last owner"
-                logger.warning(API_RESOURCE_CONFLICT, reason=msg)
-                raise ConflictError(msg) from exc
-            logger.error(
-                API_USER_SAVE_FAILED,
-                user_id=user.id,
-                intent="grant_org_role",
-                role=data.role.value,
-                constraint=exc.constraint,
-            )
-            raise
-        except QueryError:
-            logger.error(
-                API_USER_SAVE_FAILED,
-                user_id=user.id,
-                intent="grant_org_role",
-                role=data.role.value,
-            )
-            raise
-        logger.info(
-            SECURITY_PERMISSION_GRANTED,
-            user_id=user.id,
-            role=data.role.value,
-            scoped_departments=tuple(data.scoped_departments),
-        )
-        return ApiResponse(data=_to_response(updated))
-
-    @delete(
-        "/{user_id:str}/org-roles/{role:str}",
-        status_code=HTTP_204_NO_CONTENT,
-        guards=[
-            per_op_rate_limit_from_policy("users.revoke_org_role", key="user"),
-        ],
-    )
-    async def revoke_org_role(
-        self,
-        state: State,
-        user_id: PathId,
-        role: PathName,
-    ) -> None:
-        """Revoke an org-level role from a user.
-
-        Args:
-            state: Application state.
-            user_id: Target user identifier.
-            role: OrgRole value to revoke.
-
-        Raises:
-            NotFoundError: If the user is not found.
-            ValidationError: If the role value is invalid.
-            ConflictError: If revoking the last owner.
-            ConstraintViolationError: Raised on the corresponding failure path.
-            QueryError: Raised on the corresponding failure path.
-        """
-        service = _service(state)
-        try:
-            org_role = OrgRole(role)
-        except ValueError:
-            msg = f"Invalid org role: {role}"
-            logger.warning(API_VALIDATION_FAILED, reason=msg)
-            raise ValidationError(msg) from None
-
-        user = await _get_user_or_404(service, user_id, operation="revoke_org_role")
-
-        if org_role not in user.org_roles:
-            msg = f"User does not have role: {role}"
-            logger.warning(API_RESOURCE_NOT_FOUND, reason=msg)
-            raise NotFoundError(msg)
-
-        new_roles = tuple(r for r in user.org_roles if r != org_role)
-        now = datetime.now(UTC)
-        updated = user.model_copy(
-            update={
-                "org_roles": new_roles,
-                "scoped_departments": ()
-                if org_role == OrgRole.DEPARTMENT_ADMIN
-                else user.scoped_departments,
-                "updated_at": now,
-            },
-        )
-        try:
-            await service.save_update(
-                updated,
-                intent="revoke_org_role",
-                revoked_org_role=role,
-            )
-        except ConstraintViolationError as exc:
-            if exc.constraint == LAST_OWNER_TRIGGER:
-                msg = "Cannot revoke the last owner role"
-                logger.warning(API_RESOURCE_CONFLICT, reason=msg)
-                raise ConflictError(msg) from exc
-            logger.error(
-                API_USER_SAVE_FAILED,
-                user_id=user.id,
-                intent="revoke_org_role",
-                role=role,
-                constraint=exc.constraint,
-            )
-            raise
-        except QueryError:
-            logger.error(
-                API_USER_SAVE_FAILED,
-                user_id=user.id,
-                intent="revoke_org_role",
-                role=role,
-            )
-            raise
-        logger.info(
-            SECURITY_PERMISSION_REVOKED,
-            user_id=user.id,
-            role=role,
-        )
